refactor(semantic): split backend helpers
Move FAISS artifact and row loading logic plus pgvector row and rerank logic into focused modules.

Backend class imports stay stable, so existing semantic backend selection behavior is unchanged.
